+++ src/index.ts
@@ -1,9 +1,8 @@
 import { Plugin } from "@yarnpkg/core";
 import "./authHook";
-import { VersionCommand } from "./commands/version";
 import { MigrateCommand } from "./commands/migrate";
 
 const plugin: Plugin = {
-  commands: [VersionCommand, MigrateCommand],
+  commands: [MigrateCommand],
 };
 export default plugin;